Backport PR #3683: (chore): extract verbosity
