+++ src/scanpy/_settings/__init__.py
@@ -2,23 +2,23 @@ from __future__ import annotations
 
 import inspect
 import sys
-from contextlib import contextmanager
-from enum import EnumMeta, IntEnum
 from functools import wraps
-from logging import getLevelName
 from pathlib import Path
 from time import time
 from typing import TYPE_CHECKING, Literal, ParamSpec, TypeVar, get_args
 
-from . import logging
-from ._compat import deprecated, old_positionals
-from ._singleton import SingletonMeta
-from .logging import _RootLogger, _set_log_file, _set_log_level
+from .. import logging
+from .._compat import old_positionals
+from .._singleton import SingletonMeta
+from ..logging import _RootLogger, _set_log_file, _set_log_level
+from .verbosity import Verbosity
 
 if TYPE_CHECKING:
-    from collections.abc import Callable, Generator, Iterable
+    from collections.abc import Callable, Iterable
     from types import UnionType
-    from typing import ClassVar, Concatenate, LiteralString, Self, TextIO
+    from typing import ClassVar, Concatenate, Self, TextIO
+
+    from .verbosity import _VerbosityName
 
     # Collected from the print_* functions in matplotlib.backends
     _Format = (
@@ -26,15 +26,8 @@ if TYPE_CHECKING:
         | Literal["pdf", "ps", "eps", "svg", "svgz", "pgf"]
         | Literal["raw", "rgba"]
     )
-    _VerbosityName = Literal["error", "warning", "info", "hint", "debug"]
-    _LoggingLevelName = Literal["CRITICAL", "ERROR", "WARNING", "INFO", "HINT", "DEBUG"]
-else:
-    try:
-        from typing import LiteralString
-    except ImportError:
-        LiteralString = str
 
-L = TypeVar("L", bound=LiteralString)
+
 S = TypeVar("S")
 T = TypeVar("T")
 P = ParamSpec("P")
@@ -42,71 +35,6 @@ R = TypeVar("R")
 
 
 AnnDataFileFormat = Literal["h5ad", "zarr"]
-
-
-_VERBOSITY_TO_LOGLEVEL: dict[int | str, _LoggingLevelName] = {
-    "error": "ERROR",
-    "warning": "WARNING",
-    "info": "INFO",
-    "hint": "HINT",
-    "debug": "DEBUG",
-}
-_VERBOSITY_TO_LOGLEVEL.update(dict(enumerate(list(_VERBOSITY_TO_LOGLEVEL.values()))))
-
-
-class VerbosityMeta(EnumMeta):
-    @property
-    @deprecated("Use `Verbosity.warning` instead")
-    def warn(cls) -> Verbosity:
-        return Verbosity.warning
-
-
-class Verbosity(IntEnum, metaclass=VerbosityMeta):
-    """Logging verbosity levels for :attr:`scanpy.settings.verbosity`."""
-
-    error = 0
-    """Error (`0`)"""
-    warning = 1
-    """Warning (`1`)"""
-    info = 2
-    """Info (`2`)"""
-    hint = 3
-    """Hint (`3`)"""
-    debug = 4
-    """Debug (`4`)"""
-
-    def __eq__(self, other: Verbosity | int | str) -> bool:
-        if isinstance(other, Verbosity):
-            return self is other
-        if isinstance(other, int):
-            return self.value == other
-        if isinstance(other, str):
-            return self.name == other
-        return NotImplemented
-
-    @property
-    def level(self) -> int:
-        """The :ref:`logging level <levels>` corresponding to this verbosity level."""
-        # getLevelName(str) returns the int level…
-        return getLevelName(_VERBOSITY_TO_LOGLEVEL[self.name])
-
-    @contextmanager
-    def override(
-        self, verbosity: Verbosity | str | int
-    ) -> Generator[Verbosity, None, None]:
-        """Temporarily override verbosity.
-
-        >>> import scanpy as sc
-        >>> sc.settings.verbosity = sc.Verbosity.info
-        >>> with sc.settings.verbosity.override(settings.verbosity.debug):
-        ...     sc.settings.verbosity
-        <Verbosity.debug: 4>
-        >>> sc.settings.verbosity
-        <Verbosity.info: 2>
-        """
-        settings.verbosity = verbosity
-        yield self
-        settings.verbosity = self
 
 
 def _type_check(var: object, name: str, types: type | UnionType):
@@ -178,25 +106,19 @@ class SettingsMeta(SingletonMeta):
 
     @verbosity.setter
     def verbosity(cls, verbosity: Verbosity | _VerbosityName | int) -> None:
-        verbosity_str_options: list[_VerbosityName] = [
-            v for v in _VERBOSITY_TO_LOGLEVEL if isinstance(v, str)
-        ]
-        if isinstance(verbosity, Verbosity):
-            cls._verbosity = verbosity
-        elif isinstance(verbosity, int):
-            cls._verbosity = Verbosity(verbosity)
-        elif isinstance(verbosity, str):
-            verbosity = verbosity.lower()
-            if verbosity not in verbosity_str_options:
-                msg = (
-                    f"Cannot set verbosity to {verbosity}. "
-                    f"Accepted string values are: {verbosity_str_options}"
-                )
-                raise ValueError(msg)
-            cls._verbosity = Verbosity(verbosity_str_options.index(verbosity))
-        else:
-            _type_check(verbosity, "verbosity", str | int)
-        _set_log_level(cls, _VERBOSITY_TO_LOGLEVEL[cls._verbosity.name])
+        try:
+            cls._verbosity = (
+                Verbosity[verbosity.lower()]
+                if isinstance(verbosity, str)
+                else Verbosity(verbosity)
+            )
+        except KeyError:
+            msg = (
+                f"Cannot set verbosity to {verbosity}. "
+                f"Accepted string values are: {Verbosity.__members__.keys()}"
+            )
+            raise ValueError(msg) from None
+        _set_log_level(cls, cls._verbosity.level)
 
     @property
     def N_PCS(cls) -> int:
@@ -387,14 +309,14 @@ class SettingsMeta(SingletonMeta):
 
     @logfile.setter
     def logfile(cls, logfile: Path | str | TextIO | None) -> None:
-        if not hasattr(logfile, "write") and logfile:
+        if not logfile:  # "" or None
+            logfile = sys.stdout if cls._is_run_from_ipython() else sys.stderr
+        if isinstance(logfile, Path | str):
             cls.logpath = logfile
-        else:  # file object
-            if not logfile:  # None or ''
-                logfile = sys.stdout if cls._is_run_from_ipython() else sys.stderr
-            cls._logfile = logfile
-            cls._logpath = None
-            _set_log_file(cls)
+            return
+        cls._logfile = logfile
+        cls._logpath = None
+        _set_log_file(cls)
 
     @property
     def categories_to_ignore(cls) -> list[str]:
@@ -504,7 +426,7 @@ class SettingsMeta(SingletonMeta):
             rcParams["figure.facecolor"] = facecolor
             rcParams["axes.facecolor"] = facecolor
         if scanpy:
-            from .plotting._rcmod import set_rcParams_scanpy
+            from ..plotting._rcmod import set_rcParams_scanpy
 
             set_rcParams_scanpy(fontsize=fontsize, color_map=color_map)
         if figsize is not None:
