+++ src/scanpy/logging.py
@@ -16,7 +16,7 @@ if TYPE_CHECKING:
 
     from session_info2 import SessionInfo
 
-    from ._settings import settings
+    from ._settings import SettingsMeta
 
 
 # This is currently the only documented API
@@ -72,7 +72,7 @@ class _RootLogger(logging.RootLogger):
         return self.log(DEBUG, msg, time=time, deep=deep, extra=extra)
 
 
-def _set_log_file(settings: settings) -> None:
+def _set_log_file(settings: SettingsMeta) -> None:
     file = settings.logfile
     name = settings.logpath
     root = settings._root_logger
@@ -84,7 +84,7 @@ def _set_log_file(settings: settings) -> None:
     root.addHandler(h)
 
 
-def _set_log_level(settings: settings, level: int) -> None:
+def _set_log_level(settings: SettingsMeta, level: int) -> None:
     root = settings._root_logger
     root.setLevel(level)
     for h in list(root.handlers):
